Refactoring
I've taken some time to refactor your code, this is untested but generally should give you a good idea what you can do to refactor.

There's still a lot that can be done to optimize but this refactoring will allow those steps to be taken with less hassle.
